requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dj-rest-auth==7.1.1
Django==6.0.1
django-allauth==65.14.3
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
drf_spectacular==0.29.0
django-filter==25.2
channels==4.2.2
channels-redis==4.2.1